+++ src/functions/mod.rs
@@ -1,2 +1,3 @@
+pub mod rdf_bulk;
 pub mod rdf_triple;
 pub mod sparql_query;